requirement.txt
asgiref==3.2.10
certifi==2022.9.14
charset-normalizer==2.1.1
Django==3.1.2
django-filter==22.1
djangorestframework==3.13.1
idna==3.4
importlib-metadata==4.12.0
Markdown==3.4.1
psycopg2==2.9.3
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
urllib3==1.26.12
zipp==3.8.1
